norm check, wildcard pointer and pattern mathching issue
