+++ include/linked_list.h
@@ -6,7 +6,7 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 20:23:20 by jim               #+#    #+#             */
-/*   Updated: 2022/07/26 14:56:15 by jim              ###   ########seoul.kr  */
+/*   Updated: 2022/08/08 20:07:21 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,14 +20,6 @@ typedef struct s_LinkedListNode
 	void					*node;
 }				t_LinkedListNode;
 
-/*
-typedef struct s_parse_node
-{
-	char			*token_val;
-	t_token_type	type;
-}				t_parse_node;
-*/
-
 typedef struct s_LinkedList
 {
 	t_LinkedListNode	*header_node;
@@ -40,25 +32,10 @@ typedef struct s_list
 	struct s_list	*next;
 }				t_list;
 
-/*
-	for builtin test
-*/
-// typedef struct	cm
-
-t_LinkedList		*create_linked_list(void);
-// t_parse_node		*create_parse_node(char *token_val, t_token_type type);
-// t_LinkedListNode	*create_linked_node(char	*str, t_token_type type);
-t_LinkedListNode	*get_linked_node(t_LinkedList *pLinkedList, \
-											char *data_str);
-int					add_back_linked_node(t_LinkedList *pLinkedList, \
-										t_LinkedListNode *newLinkedListNode);
-int					remove_linked_node(t_LinkedList *pLinkedList, \
-											char *data_str);
-int					clear_linked_list(t_LinkedList *pLinkedList);
-int					delete_linked_list(t_LinkedList **PLinkedList);
-int					display_linked_list(t_LinkedList *pLinkedList);
-
-void				free_linked_list(t_list **list);
-t_list				*create_list(void);
+void	free_node(t_list **list);
+void	safe_free_token(t_list **token);
+void	safe_free_token_list(t_list *list);
+int		wrapper_free_token_list(t_list *list, int return_val);
+int		wrapper_free_list(char ***word_list);
 
 #endif
