+++ srcs/builtin/export.c
@@ -6,7 +6,7 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/29 18:19:01 by jim               #+#    #+#             */
-/*   Updated: 2022/08/08 13:54:05 by jim              ###   ########.fr       */
+/*   Updated: 2022/08/08 14:29:33 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,6 +47,29 @@ static t_env_node	*find_export_key(t_env_list *env_list, char *key)
 	return (NULL);
 }
 
+static int	export_result_replace_or_add(t_env_list *env_list, \
+											char *key, char *value)
+{
+	t_env_node	*found_node;
+	t_env_node	*new_node;
+
+	found_node = find_export_key(env_list, key);
+	if (found_node != NULL && value != NULL)
+	{
+		if (found_node->value != NULL)
+			free(found_node->value);
+		found_node->value = value;
+	}
+	else if (found_node == NULL)
+	{
+		new_node = create_env_node(key, value);
+		if (new_node == NULL)
+			return (-1);
+		add_back_env_node(env_list, new_node);
+	}
+	return (0);
+}
+
 /*
 	- "="이 있는지 확인한다.
 	- 없으면 해당 인자를 key로만 넣고 value를  NULL로 처리힌다.
@@ -60,8 +83,6 @@ export a b c=42같은 케이스 처리되게 해야한다.
 */
 static int	add_to_export_list(t_env_list *env_list, char **arg_list)
 {
-	t_env_node	*new_node;
-	t_env_node	*found_node;
 	size_t		idx;
 	char		*key;
 	char		*value;
@@ -77,20 +98,8 @@ static int	add_to_export_list(t_env_list *env_list, char **arg_list)
 			idx++;
 			continue ;
 		}
-		found_node = find_export_key(env_list, key);
-		if (found_node != NULL && value != NULL)
-		{
-			if (found_node->value != NULL)
-				free(found_node->value);
-			found_node->value = value;
-		}
-		else if (found_node == NULL)
-		{
-			new_node = create_env_node(key, value);
-			if (new_node == NULL)
-				return (-1);
-			add_back_env_node(env_list, new_node);
-		}
+		if (export_result_replace_or_add(env_list, key, value) < 0)
+			return (-1);
 		idx++;
 	}
 	return (0);
