+++ srcs/data_structure/linked_list/env_list_remove.c
@@ -6,7 +6,7 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 20:23:01 by jim               #+#    #+#             */
-/*   Updated: 2022/07/04 11:44:51 by jim              ###   ########.fr       */
+/*   Updated: 2022/08/08 14:37:01 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
