+++ srcs/execute/execute_builtin.c
@@ -6,7 +6,7 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/19 18:28:18 by jim               #+#    #+#             */
-/*   Updated: 2022/07/31 11:32:47 by jim              ###   ########.fr       */
+/*   Updated: 2022/08/08 14:58:40 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,22 +37,15 @@ int	check_builtin(t_list *cmd_list)
 	return (0);
 }
 
-/* 25줄 */
-int builtin_process(t_env_list *env_list, t_list *cmd_list, int is_child)
+int	builtin_process(t_env_list *env_list, char **cmd, int is_child)
 {
 	int		status;
 	char	**argv;
-	char	**cmd;
 
-	if (cmd_list == NULL)
-		return (0);
-	cmd = list_to_array(cmd_list);
-	if (cmd == NULL)
-		return (2);
-    status = 0;
-    argv = &(cmd[1]);
-    if (ft_strncmp("exit", *cmd, max_nonnegative("exit", *cmd)) == 0)
-		exit_cmd(argv, is_child); // exit의 경우 고민 필요 
+	status = 0;
+	argv = &(cmd[1]);
+	if (ft_strncmp("exit", *cmd, max_nonnegative("exit", *cmd)) == 0)
+		exit_cmd(argv, is_child);
 	else if (ft_strncmp("echo", *cmd, max_nonnegative("echo", *cmd)) == 0)
 		status = echo_cmd(argv);
 	else if (ft_strncmp("cd", *cmd, max_nonnegative("cd", *cmd)) == 0)
@@ -68,5 +61,5 @@ int builtin_process(t_env_list *env_list, t_list *cmd_list, int is_child)
 	if (is_child)
 		exit(status);
 	free_list(&cmd);
-    return (status);
+	return (status);
 }
