+++ srcs/execute/execute_utils.c
@@ -6,7 +6,7 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/03 16:32:30 by jim               #+#    #+#             */
-/*   Updated: 2022/08/03 16:37:22 by jim              ###   ########.fr       */
+/*   Updated: 2022/08/08 15:00:09 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ static int	safe_strjoin(char **dst, char *s1, char *s2, char **depend_list)
 	return (0);
 }
 
-static int get_length(t_list *plist)
+static int	get_length(t_list *plist)
 {
 	t_list	*cur_node;
 	int		list_len;
@@ -66,7 +66,6 @@ char	**list_to_array(t_list *plist)
 	arr_list[list_len] = NULL;
 	return (arr_list);
 }
-/**/
 
 char	**env_list_to_array(t_env_list *env_list)
 {
