+++ srcs/minishell.c
@@ -6,7 +6,7 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/24 22:49:58 by jim               #+#    #+#             */
-/*   Updated: 2022/08/04 15:15:12 by jim              ###   ########seoul.kr  */
+/*   Updated: 2022/08/08 16:25:00 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@
 #include "lexer.h"
 #include "heredoc.h"
 #include "ft_signal.h"
-//debug
-#include <stdlib.h>
 #include "linked_list.h"
+#include <stdlib.h>
+//debug
 #include <string.h>
 #include <sys/errno.h>
 
@@ -42,26 +42,9 @@ static int	preprocess(char *input, t_list *parsed_header)
 	if (!parser(token_header.next, parsed_header))
 	{
 		printf("%s\n", "parser error");
-		// free(token_header);
-		// free(parsed_header);
 		return (2);
 	}
-	// heredoc_processing();
 	return (0);
-}
-
-// for display, will be removed
-static void	display_env_list(t_env_list	*env_list)
-{
-	t_env_node	*cur_node;
-
-	cur_node = env_list->header_node;
-	while (cur_node)
-	{
-		printf("cur_node.key : %s\n", cur_node->key);
-		printf("cur_node.value : %s\n", cur_node->value);
-		cur_node = cur_node->next_node;
-	}
 }
 
 static int	reset_in_out_fd(int io_backup[2])
@@ -101,7 +84,6 @@ int	main(int argc, char **argv, char **envp)
 			continue ;
 		if (preprocess(input, &parsed_header))
 			continue ;
-		// print_command_content(parsed_header.next);
 		execute_processing(env_list, parsed_header.next, FALSE, &parsed_header);
 		add_history(input);
 		// error 발생시 free시키는 조건을 일괄적으로 할 필요가 있다.
