+++ srcs/utils/check_valid_variable.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   check_valid_variable.c                                 :+:      :+:    :+:   */
+/*   check_valid_variable.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/06 12:14:21 by jim               #+#    #+#             */
-/*   Updated: 2022/07/30 18:44:22 by jim              ###   ########.fr       */
+/*   Updated: 2022/08/08 15:33:41 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
