+++ srcs/utils/exit_status.c
@@ -6,14 +6,14 @@
 /*   By: jim <jim@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/06 16:06:09 by jim               #+#    #+#             */
-/*   Updated: 2022/08/03 15:35:56 by jim              ###   ########.fr       */
+/*   Updated: 2022/08/08 16:12:10 by jim              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "linked_list.h"
 #include "env_list.h"
 
-int	*get_exit_status()
+int	*get_exit_status(void)
 {
 	static int	exit_status;
 
@@ -24,7 +24,7 @@ void	update_exit_status(int status, t_list *org_list)
 {
 	if (status < 0)
 	{
-		// delete_list(&org_list)
+		// delete_list(&org_list);
 		status = 2;
 	}
 	*(get_exit_status()) = status;
