+++ t2.txt
@@ -1,0 +1,1 @@
+aabdf
